Peak ugen improvements
this patch adds support for a control-rate Peak ugen with audio-rate
input.

code like:
Peak.kr(In.ar(0))
is not forbidden any more, but the peak value of the input block is
computed and sent to the output in control rate.

with this feature, it would is possible to implement efficient envelope
followers e.g. for vu meter guis.

git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@9291 a380766d-ff14-0410-b294-a243070f3f08