sclang: ignore SIGCHLD
Jakob Leben:

When testing QtCollider, I realized that all communication from external
world to sclang through pipes was lost after scsynth has been started as a
child process of sclang, and then stopped. The reason was that the thread in
which SC_TerminalClient was running quitted. By investigating further I
found out the following:

When a QApplication is started, it installs a handler on SIGCHLD signal,
thus changing the default behaviour that the signal is ignored.

Effectively, when scsynth (that has been started from sclang as a child
process) finishes, it sends SIGCHLD to sclang, which interrupts the poll()
system call in SC_TerminalClient::commandLoop(). Poll() then returns -1, in
which case SC_TerminalClient blindly quit()s the thread, without checking
the errno, which has been set to EINTR by poll().

From what I've read, it is common, contrary to supercollider's behaviour, to
just retry a system call (in this case poll() ) after it is interrupted. I
tested this in supercollider and everything seems ok so far.

Signed-off-by: Tim Blechmann <tim@klingt.org>

git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@10294 a380766d-ff14-0410-b294-a243070f3f08